Added README Template.
The main goal of the file is to make using it with a real project mostly a process of removing template sections that aren't needed. Because of that, there's a lot of extra stuff that won't be applicable to all projects. This is also to provide hints to developers about WHAT to document.

Signed-off-by: Brian Porter <beporter@users.sourceforge.net>