WIP. Non-functional.
So the basic idea here is that instead of binding a hidden `grunt.even.on('watch)` event, we could register a multiTask directly, and loop over the supplied filenames.

The problem seems to be that grunt-contrib-watch doesn't expect to hand off to a multiTask, so things don't work correctly. There seems to be an approach for handling this kind of case laid out here though: https://www.npmjs.com/package/grunt-contrib-watch#compiling-files-as-needed

Needs more experimentation.