Customizes puphpet config, Vagrantfile and manifests.
Imports important changes from the 2.x skeleton.

Also brings in a much more "robust" VM config, but following the rule that it's easier to delete things than add, a reasonably complete set of packages is included for installation into the VM. If a project wants faster provisioning and doesn't need the extra deps, they can be safely removed as needed.

This also lets the skeleton act as a handy reference for our preferred "addon" packages when a project needs to add additional features back in later on.

Also makes the config file templated and includes minimal tokens.