Install new "simple" Vagrantfile and shell provisioning.
Taken from my ENvAwareness demo app, this is a slightly more sophisticated approach that breaks out the provisioning shell script into an actual .sh file, along with the Apache vhost.conf file and an .sql file containing the commands to run.

Basic work is complete to make this setup env-aware, allowing the shell script to be run on a "real" server as well as in vagrant. It sets the appropriate value for the APP_ENV environment variable into a new common source file, `/etc/app_env`, which is sourced by both the `vagrant` user's `~/.profile` and Apache's `/etc/apache2/envvars` file.