More FileParser tests.
There are issues with the token replacements-- it happens using `str_replace()` which requires the **exact** matching token string. This won't work if the token is used in more than one file and has different default values in each one since we only preserve a single duplicate.

A better approach would be to still use `preg_replace_all()` and craft the regex to match using the token name and an open-ended default value matcher.

Or perhaps use a callback. We could match ALL tokens and then use the submatch of the token name to determine the proper replacement value.

Matched tokens should probably also be objects that can store multiple properties:

* token name
* array of defaults (per file)
* user-entered replacement value

It'd be cleaner to associate methods for fetching this data, as well as generating matchers/replacers, from that class.