Removed /*position: fixed;*/ from div.message{ }
This made the flash notification message to always spread atop the contents of the page instead of being squeezed to the top right corner.
The way it was before i changed it, it was covering the links at the top right corner of the page such as 'documentation' and 'API' links.

I also added
text-align: center; just to make sure the text stays in the center of the div.

Signed-off-by: Rick Guyer <ricoguyer@gmail.com>