WIP script to provision necessary resources on AWS programmatically.
The thought with this work is that it would dovetail with the `bootstrap.sh` script, allowing Loadsys to programmatically build a "standard" hosting environment in AWS that was compatible with our app skeleton out of the box. A configuration file would be used to store "answers" to questions about naming, created instance IDs, etc.